+++ Assets/Scripts/Prototype/Note/Note.cs
@@ -24,7 +24,7 @@ namespace Plutono.Song
         public void FallingDown()
         {
             //Force the note to back to the original position where it begins to fall down
-            gameObject.transform.position = new Vector3(_details.pos * 10, 0, Settings.maximumNoteRange);
+            gameObject.transform.position = new Vector3((float)(_details.pos * 10), 0, Settings.maximumNoteRange);
             gameObject.transform.DOMoveZ(0, Settings.NoteFallTime(10)).SetEase(Ease.Linear);
         }
